Revert "[Refactor] Reduce size of position and undo structure for easier copying of splitpoint data in SMP."
This reverts commit 70a167c9f7103678bcfec3fee0cb72ac68386287.

+++ typedefs.h
@@ -93,12 +93,12 @@ struct movelist_t {
 struct pos_store_t {
     pos_store_t() : previous(nullptr) {}
     uint32 lastmove;
-    int8 castle;
-    int8 fifty;
-    int8 epsq;
-    int8 pliesFromNull;
+    int castle;
+    int fifty;
+    int epsq;
+    int pliesFromNull;
     EvalScore posScore[2];
-    int16 mat_summ[2];
+    int mat_summ[2];
     uint64 phash;
     uint64 hash;
     pos_store_t* previous;
@@ -114,10 +114,10 @@ struct position_t {
     uint64 kings;
     uint64 color[2];
     uint64 occupied;
-    int8 pieces[64];
-    int8 kpos[2];
+    int pieces[64];
+    int kpos[2];
     pos_store_t posStore;
-    int8 side;
+    int side;
 };
 
 typedef uint8 mflag_t;
